Remove function prop from AcceptInvitationButtons
