+++ src/client/pages/Login/components/AcceptInvitationButtons/AcceptInvitationButtons.tsx
@@ -1,32 +1,50 @@
 import React from 'react'
 import { useTranslation } from 'react-i18next'
+import { useNavigate } from 'react-router-dom'
 
 import { ApiEndPoint } from 'meta/api/endpoint'
-import { LoginInvitationQueryParams } from 'meta/routes'
+import { LoginInvitationQueryParams, Routes } from 'meta/routes'
+import { AuthProvider } from 'meta/user'
 
 import { useInvitation } from 'client/store/login'
+import { useAcceptInvitationForm } from 'client/store/login/hooks'
+import { useIsInvitationLocalRoute } from 'client/hooks/useIsRoute'
 import { useSearchParams } from 'client/hooks/useSearchParams'
 import Icon from 'client/components/Icon'
 import { videoResources } from 'client/pages/Tutorials'
 
-type Props = {
-  onAcceptInvitationLocalClick: () => void
-}
+import { useOnAcceptInvitationLocal } from './hooks/useOnAcceptInvitationLocal'
 
-const AcceptInvitationButtons: React.FC<Props> = (props) => {
-  const { onAcceptInvitationLocalClick } = props
+const AcceptInvitationButtons: React.FC = () => {
   const { i18n, t } = useTranslation()
+  const navigate = useNavigate()
 
-  const { invitationUuid } = useSearchParams<LoginInvitationQueryParams>()
-  const { userRole, assessment } = useInvitation()
+  const { invitationUuid, lang } = useSearchParams<LoginInvitationQueryParams>()
+  const { assessment, invitedUser, userProviders, userRole } = useInvitation()
 
   const cycle = assessment?.cycles.find((cycle) => cycle.uuid === userRole.cycleUuid)
   const assessmentName = assessment?.props.name
   const cycleName = cycle?.name
 
+  const isInInvitationLocal = useIsInvitationLocalRoute()
+
+  const formData = useAcceptInvitationForm()
+  const showPassword2 =
+    (invitedUser && !userProviders) || (userProviders && !userProviders.includes(AuthProvider.local))
+
+  const onAcceptInvitationLocal = useOnAcceptInvitationLocal({ formData, invitationUuid, showPassword2 })
+
+  const goToAcceptInvitationLocal = () => {
+    navigate(Routes.LoginInvitationLocal.generatePath({ assessmentName, cycleName }, { invitationUuid, lang }))
+  }
+
   return (
     <>
-      <button type="button" className="btn" onClick={onAcceptInvitationLocalClick}>
+      <button
+        type="button"
+        className="btn"
+        onClick={isInInvitationLocal ? onAcceptInvitationLocal : goToAcceptInvitationLocal}
+      >
         {t('login.acceptInvitationWithFra')}
       </button>
